Renamed db/nagios_status_data.py to db/nagios_config_wrapper.py and function within, and added another unittest
